sprit to reporter
