+++ planning/autoware_trajectory_validator/src/filters/safety/collision_check_filter/reporter.cpp
@@ -14,16 +14,33 @@
 
 #include "reporter.hpp"
 
+#include <autoware_internal_planning_msgs/msg/control_point.hpp>
+#include <autoware_internal_planning_msgs/msg/planning_factor.hpp>
+#include <autoware_internal_planning_msgs/msg/safety_factor.hpp>
+#include <autoware_internal_planning_msgs/msg/safety_factor_array.hpp>
 #include <autoware_trajectory_validator/msg/metric_report.hpp>
 #include <geometry_msgs/msg/point.hpp>
 #include <rclcpp/logging.hpp>
 
+#include <fmt/core.h>
+
+#include <cstdint>
+#include <string>
 #include <utility>
+#include <vector>
 
 namespace autoware::trajectory_validator::plugin::safety::reporter
 {
 namespace
 {
+using autoware_trajectory_validator::msg::MetricReport;
+
+struct VisualizationData
+{
+  std::string error_msg{};
+  autoware_internal_planning_msgs::msg::PlanningFactorArray planning_factors{};
+};
+
 struct Color
 {
   float r;
@@ -45,6 +62,194 @@ Color resolve_trajectory_color(const std::string & trajectory_id)
     return Color{0.0F, 0.75F, 1.0F};
   }
   return Color{0.2F, 1.0F, 0.2F};
+}
+
+autoware_internal_planning_msgs::msg::SafetyFactorArray make_safety_factor_array(
+  const builtin_interfaces::msg::Time & stamp, const CollisionDetail & collision_detail,
+  const std::string & collision_type, double time_resolution)
+{
+  using autoware_internal_planning_msgs::msg::SafetyFactor;
+  using autoware_internal_planning_msgs::msg::SafetyFactorArray;
+
+  SafetyFactor safety_factor;
+  safety_factor.type = SafetyFactor::OBJECT;
+  safety_factor.object_id = collision_detail.object_identification.uuid;
+  safety_factor.ttc_begin = static_cast<float>(collision_detail.ttc);
+  safety_factor.ttc_end = static_cast<float>(collision_detail.ttc + time_resolution);
+  safety_factor.is_safe = false;
+  if (!collision_detail.object_trajectory.empty()) {
+    safety_factor.points.push_back(collision_detail.object_trajectory.front().position);
+  }
+
+  SafetyFactorArray safety_factors;
+  safety_factors.header.stamp = stamp;
+  safety_factors.header.frame_id = "map";
+  safety_factors.factors.push_back(std::move(safety_factor));
+  safety_factors.is_safe = false;
+  safety_factors.detail = collision_type;
+  return safety_factors;
+}
+
+void add_collision_planning_factor(
+  const double time_resolution, const builtin_interfaces::msg::Time & stamp,
+  const geometry_msgs::msg::Pose & ego_pose, const CollisionDetail & collision_detail,
+  const std::string & collision_type,
+  autoware_internal_planning_msgs::msg::PlanningFactorArray & planning_factors)
+{
+  const auto safety_factors =
+    make_safety_factor_array(stamp, collision_detail, collision_type, time_resolution);
+  const auto control_point =
+    autoware_internal_planning_msgs::build<autoware_internal_planning_msgs::msg::ControlPoint>()
+      .pose(ego_pose)
+      .velocity(0.0)
+      .shift_length(0.0)
+      .distance(0.0);
+  auto factor =
+    autoware_internal_planning_msgs::build<autoware_internal_planning_msgs::msg::PlanningFactor>()
+      .module("")
+      .is_driving_forward(true)
+      .control_points({control_point})
+      .behavior(autoware_internal_planning_msgs::msg::PlanningFactor::STOP)
+      .detail(collision_type)
+      .safety_factors(safety_factors);
+  planning_factors.factors.push_back(std::move(factor));
+}
+
+void process_pet_artifacts(
+  const nav_msgs::msg::Odometry & odometry,
+  reporter::ContinuousDetectionTimes & pet_continuous_times, const rclcpp::Time & current_time,
+  const PetArtifact & pet_artifact, VisualizationData & artifacts,
+  visualization_msgs::msg::MarkerArray & debug_markers, double time_resolution)
+{
+  pet_continuous_times.update(
+    current_time, pet_artifact.object_evaluations, [](const auto & evaluation) {
+      return evaluation.detail.object_identification.trajectory_id_string();
+    });
+
+  if (pet_artifact.risk == RiskLevel::SAFE || pet_artifact.object_evaluations.empty()) {
+    return;
+  }
+
+  std::string log_messages{};
+  std::string marker_messages{};
+  uint8_t log_level = MetricReport::WARN;
+  for (const auto & evaluation : pet_artifact.object_evaluations) {
+    if (evaluation.risk == RiskLevel::SAFE) {
+      continue;
+    }
+
+    const auto & timing = evaluation.detail;
+    const auto & obj_id = timing.object_identification;
+    const bool is_error = evaluation.risk == RiskLevel::ERROR;
+    if (is_error) {
+      log_level = MetricReport::ERROR;
+    }
+
+    const auto finding_msg = fmt::format(
+      "PET collision, classification: {}, ID: {}, PET: {}, TTC: {}, duration: {}, stamp: {}.{};",
+      obj_id.classification, obj_id.trajectory_id_string(), timing.pet, timing.ttc,
+      pet_continuous_times.get_time(obj_id.trajectory_id_string()), obj_id.stamp.sec,
+      obj_id.stamp.nanosec);
+    log_messages += finding_msg;
+    reporter::append_text_marker_message(marker_messages, finding_msg);
+    reporter::add_debug_markers(
+      debug_markers, current_time, "planned_speed_collision", obj_id.trajectory_id_string(),
+      timing.ego_trajectory, timing.object_trajectory, timing.ego_hull, timing.object_hull);
+    if (is_error) {
+      add_collision_planning_factor(
+        time_resolution, odometry.header.stamp, odometry.pose.pose, timing, "PET",
+        artifacts.planning_factors);
+    }
+  }
+
+  artifacts.error_msg += marker_messages;
+  reporter::log_collision_messages(log_level, log_messages);
+}
+
+void process_drac_artifacts(
+  const nav_msgs::msg::Odometry & odometry,
+  reporter::ContinuousDetectionTimes & drac_continuous_times, const rclcpp::Time & current_time,
+  const DracArtifact & drac_artifact, VisualizationData & artifacts,
+  visualization_msgs::msg::MarkerArray & debug_markers, double time_resolution)
+{
+  drac_continuous_times.update(
+    current_time, drac_artifact.object_evaluations, [](const auto & evaluation) {
+      return evaluation.detail.object_identification.trajectory_id_string();
+    });
+
+  if (drac_artifact.risk == RiskLevel::SAFE || drac_artifact.object_evaluations.empty()) {
+    return;
+  }
+
+  std::string log_messages{};
+  std::string marker_messages{};
+  const bool has_error = drac_artifact.risk == RiskLevel::ERROR;
+  const uint8_t log_level = has_error ? MetricReport::ERROR : MetricReport::WARN;
+  for (const auto & evaluation : drac_artifact.object_evaluations) {
+    const auto & timing = evaluation.detail;
+    const auto & obj_id = timing.object_identification;
+
+    const auto finding_msg = fmt::format(
+      "DRAC collision, ID: {}, PET: {}, TTC: {}, DRAC: {}, stamp: {}.{};",
+      obj_id.trajectory_id_string(), timing.pet, timing.ttc,
+      drac_artifact.required_acceleration.has_value()
+        ? std::to_string(drac_artifact.required_acceleration.value())
+        : "Cant be avoided",
+      obj_id.stamp.sec, obj_id.stamp.nanosec);
+    log_messages += finding_msg;
+    reporter::append_text_marker_message(marker_messages, finding_msg);
+    reporter::add_debug_markers(
+      debug_markers, current_time, "drac_collision", obj_id.trajectory_id_string(),
+      timing.ego_trajectory, timing.object_trajectory, timing.ego_hull, timing.object_hull);
+    if (has_error) {
+      add_collision_planning_factor(
+        time_resolution, odometry.header.stamp, odometry.pose.pose, timing, "DRAC",
+        artifacts.planning_factors);
+    }
+  }
+
+  artifacts.error_msg += marker_messages;
+  reporter::log_collision_messages(log_level, log_messages);
+}
+
+void process_rss_artifacts(
+  const RssArtifact & rss_artifact, reporter::ContinuousDetectionTimes & rss_continuous_times,
+  const rclcpp::Time & current_time, VisualizationData & artifacts)
+{
+  std::vector<RssEvaluation> violations{};
+  violations.reserve(rss_artifact.object_evaluations.size());
+  for (const auto & evaluation : rss_artifact.object_evaluations) {
+    if (evaluation.risk == RiskLevel::SAFE) {
+      continue;
+    }
+    violations.push_back(evaluation);
+  }
+  rss_continuous_times.update(current_time, violations, [](const auto & violation) {
+    return violation.detail.object_identification.object_id_string();
+  });
+
+  if (rss_artifact.risk == RiskLevel::SAFE || violations.empty()) {
+    return;
+  }
+
+  std::string log_messages{};
+  std::string marker_messages{};
+  for (const auto & violation : violations) {
+    const auto & detail = violation.detail;
+    const auto object_id = detail.object_identification.object_id_string();
+
+    const auto finding_msg = fmt::format(
+      "RSS collision, classification: {}, ID: {}, duration: {}, required deceleration: {}, "
+      "stamp: {}.{};",
+      detail.object_identification.classification, object_id,
+      rss_continuous_times.get_time(object_id), detail.rss_acceleration,
+      detail.object_identification.stamp.sec, detail.object_identification.stamp.nanosec);
+    log_messages += finding_msg;
+    reporter::append_text_marker_message(marker_messages, finding_msg);
+  }
+
+  artifacts.error_msg += marker_messages;
+  reporter::log_collision_messages(MetricReport::ERROR, log_messages);
 }
 }  // namespace
 
@@ -190,5 +395,32 @@ void log_collision_messages(const uint8_t level, const std::string & messages)
     return;
   }
   RCLCPP_WARN(rclcpp::get_logger("CollisionCheckFilter"), "Warning: %s", messages.c_str());
+}
+
+autoware_internal_planning_msgs::msg::PlanningFactorArray process_collision_artifacts(
+  const nav_msgs::msg::Odometry & odometry, const PetArtifact & pet_artifact,
+  ContinuousDetectionTimes & pet_continuous_times, const DracArtifact & drac_artifact,
+  ContinuousDetectionTimes & drac_continuous_times, const RssArtifact & rss_artifact,
+  ContinuousDetectionTimes & rss_continuous_times,
+  visualization_msgs::msg::MarkerArray & debug_markers, double time_resolution)
+{
+  VisualizationData visualization_data{};
+  const auto current_time = rclcpp::Time{odometry.header.stamp};
+
+  process_pet_artifacts(
+    odometry, pet_continuous_times, current_time, pet_artifact, visualization_data,
+    debug_markers, time_resolution);
+  process_drac_artifacts(
+    odometry, drac_continuous_times, current_time, drac_artifact, visualization_data,
+    debug_markers, time_resolution);
+  process_rss_artifacts(
+    rss_artifact, rss_continuous_times, current_time, visualization_data);
+
+  if (!visualization_data.error_msg.empty()) {
+    add_error_text_marker(
+      debug_markers, current_time, odometry.pose.pose, visualization_data.error_msg);
+  }
+
+  return std::move(visualization_data.planning_factors);
 }
 }  // namespace autoware::trajectory_validator::plugin::safety::reporter
